vindstyrka_logger/requirements.txt
pyyaml
websockets
pyserial
